allow both int32 and int64 to be passed from the python side

+++ src/arraytools.cpp
@@ -671,7 +671,7 @@ array_link::array_link (const array_link &rhs) {
 }
 
 #ifdef SWIGCODE
-array_link::array_link (long *pymatinput, int nrows, int ncols) {
+array_link::array_link (long long *pymatinput, int nrows, int ncols) {
         this->index = INDEX_DEFAULT;
         this->n_columns = ncols;
         this->n_rows = nrows;
@@ -4637,13 +4637,13 @@ array_link::array_link (const array_t *array, rowindex_t nrows, colindex_t ncols
 }
 
 #ifdef SWIGCODE
-array_link create_array_link(long* pymatinput, int number_of_rows, int number_of_columns) {
+array_link create_array_link(long long* pymatinput, int number_of_rows, int number_of_columns) {
      array_link array = array_link(number_of_rows, number_of_columns, array_link::INDEX_DEFAULT);
      array.setarraydata_transposed(pymatinput, number_of_rows*number_of_columns);
      return array;
 }
 
-void update_array_link(array_link &al, long* pymatinput, int number_of_rows, int number_of_columns) {
+void update_array_link(array_link &al, long long* pymatinput, int number_of_rows, int number_of_columns) {
      al.init(number_of_rows, number_of_columns);
      al.setarraydata_transposed(pymatinput, number_of_rows*number_of_columns);
      return;
